requirements.txt
tensorci-client==0.0.2